Panzer: remove use of Fad version of BasisValues
-Fixes CurlLaplacian test for cuda, #3340
-Changes test to use Tpetra instead of Epetra by default
-Revert "Disable PanzerAdaptersSTK_CurlLaplacianExample-ConvTest-Quad-Order-4 on waterman CUDA (#3340)"

This reverts commit f17f9bd27fde393e92064917525998b3a3b9f1ff.
