+++ packages/panzer/disc-fe/src/evaluators/Panzer_Dirichlet_Residual_EdgeBasis.hpp
@@ -95,8 +95,8 @@ class DirichletResidual_EdgeBasis
   Teuchos::RCP<const panzer::PureBasis> basis; 
   Teuchos::RCP<const panzer::PointRule> pointRule; 
 
-  PointValues2<ScalarT> pointValues;
-  PHX::MDField<const ScalarT, Cell, IP, Dim, Dim, void, void, void, void>
+  PointValues2<double> pointValues;
+  PHX::MDField<const double, Cell, IP, Dim, Dim, void, void, void, void>
     constJac_;
 
   Teuchos::RCP<const std::vector<Intrepid2::Orientation> > orientations;
