+++ packages/panzer/disc-fe/src/evaluators/Panzer_Dirichlet_Residual_FaceBasis.hpp
@@ -97,8 +97,8 @@ class DirichletResidual_FaceBasis
   Kokkos::DynRankView<ScalarT,PHX::Device> faceNormal; // face normals
   Kokkos::DynRankView<ScalarT,PHX::Device> refFaceNormal; // reference face normals
 
-  PointValues2<ScalarT> pointValues;
-  PHX::MDField<const ScalarT, Cell, IP, Dim, Dim, void, void, void, void>
+  PointValues2<double> pointValues;
+  PHX::MDField<const double, Cell, IP, Dim, Dim, void, void, void, void>
     constJac_;
 
   Teuchos::RCP<const std::vector<Intrepid2::Orientation> > orientations;
