+++ packages/panzer/disc-fe/src/evaluators/Panzer_Dirichlet_Residual_FaceBasis_impl.hpp
@@ -86,7 +86,7 @@ DirichletResidual_FaceBasis(
   value    = PHX::MDField<const ScalarT,Cell,Point,Dim>(value_name, vector_layout_vector);
 
   // setup all fields to be evaluated and constructed
-  pointValues = PointValues2<ScalarT> (pointRule->getName()+"_",false);
+  pointValues = PointValues2<double> (pointRule->getName()+"_",false);
   pointValues.setupArrays(pointRule);
 
   // the field manager will allocate all of these field
