+++ packages/panzer/disc-fe/src/evaluators/Panzer_GatherNormals_impl.hpp
@@ -78,7 +78,7 @@ GatherNormals(
   // setup the orientation field
   std::string orientationFieldName = basis->name() + " Orientation";
   // setup all fields to be evaluated and constructed
-  pointValues = panzer::PointValues2<ScalarT> (pointRule->getName()+"_",false);
+  pointValues = panzer::PointValues2<double> (pointRule->getName()+"_",false);
   pointValues.setupArrays(pointRule);
 
   // the field manager will allocate all of these field
