+++ packages/panzer/disc-fe/src/evaluators/Panzer_PointValues_Evaluator_decl.hpp
@@ -77,7 +77,7 @@ class PointValues_Evaluator
     using ScalarT = typename EvalT::ScalarT;
 
   // is anything other than ScalarT really needed here?
-  PointValues2<ScalarT> pointValues;
+  PointValues2<double> pointValues;
  
   PHX::MDField<double,NODE,Dim> refPointArray;
